+++ src/Event/SourceFactoryCreateSourceEvent.php
@@ -8,9 +8,8 @@
 
 namespace HeimrichHannot\EntityImportBundle\Event;
 
+use Contao\Model;
 use HeimrichHannot\EntityImportBundle\Source\SourceInterface;
-use HeimrichHannot\UtilsBundle\File\FileUtil;
-use HeimrichHannot\UtilsBundle\Model\ModelUtil;
 use Symfony\Component\EventDispatcher\Event;
 
 class SourceFactoryCreateSourceEvent extends Event
@@ -18,29 +17,24 @@ class SourceFactoryCreateSourceEvent extends Event
     public const NAME = 'huh.entity_import.source_factory_create_source_event';
 
     /**
-     * @var ModelUtil
-     */
-    private $modelUtil;
-    /**
-     * @var FileUtil
-     */
-    private $fileUtil;
-    /**
      * @var SourceInterface
      */
     private $source;
+    /**
+     * @var Model
+     */
+    private $sourceModel;
 
     /**
      * ImporterFactoryCreateFileSourceEvent constructor.
      */
-    public function __construct(SourceInterface $source, FileUtil $fileUtil, ModelUtil $modelUtil)
+    public function __construct(?SourceInterface $source, Model $sourceModel)
     {
-        $this->fileUtil = $fileUtil;
-        $this->modelUtil = $modelUtil;
         $this->source = $source;
+        $this->sourceModel = $sourceModel;
     }
 
-    public function getSource(): SourceInterface
+    public function getSource(): ?SourceInterface
     {
         return $this->source;
     }
@@ -48,5 +42,10 @@ class SourceFactoryCreateSourceEvent extends Event
     public function setSource(SourceInterface $source)
     {
         $this->source = $source;
+    }
+
+    public function getSourceModel(): Model
+    {
+        return $this->sourceModel;
     }
 }